requirements.txt
pydantic
pytest
openai
python-dotenv
fastapi
uvicorn
